refactor: ♻️ #9 フォルダ構成の変更と微修正
